fix: remove .git from the repository URL
It looks like semantic-release uses the repository field to generate URLs to pull requests and commits and whatnot in the release notes, and while GitHub redirects from <project-url>.git to <project-url>, it doesn't seem to redirect sub-paths like <project-url>.git/commits/<hash>, so the links are broken. I think this will fix them.